Move wallet utility functions to the Wallet directory
