+++ src/modules/wallet/utils/api/lsk.test.js
